cleanup files for java 8 (openJdk) setup instead of java 11 (oracle)
