+++ web/.vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "java.home": "/usr/lib/jvm/java-8-openjdk-amd64"
+    "java.home": "/usr/lib/jvm/java-8-openjdk-amd64",
+    "java.configuration.updateBuildConfiguration": "automatic"
 }
